Heading + made mouse movement more UX friendly
