+++ src/plays/drawit/DrawIt.jsx
@@ -3,6 +3,7 @@ import { getPlayById } from "meta/play-meta-util";
 import PlayHeader from "common/playlists/PlayHeader";
 import "./drawIt.css";
 import { useEffect, useRef, useState } from "react";
+import Canvas from "./Canvas";
 
 function DrawIt(props) {
   // Do not remove the below lines.
@@ -11,107 +12,105 @@ function DrawIt(props) {
   const play = getPlayById(id);
 
   // Your Code Start below.
-  let canvas, context;
-  const canvasRef = useRef(null);
-  const [color, setcolor] = useState("white");
-  //   console.log(color);
+  // let canvas, context;
+  // const canvasRef = useRef(null);
+  // const [color, setcolor] = useState("white");
+  // //   console.log(color);
 
-  //   useEffect(() => {
-  // console.log("kk",color)
-  //     onColorUpdate();
+  // //   useEffect(() => {
+  // // console.log("kk",color)
+  // //     onColorUpdate();
 
-  //   }, [color]);
-  useEffect(() => {
-    // console.log("iiiiiiii", color);
-    canvas = canvasRef.current;
-    //Our first draw
-    context = canvas.getContext("2d");
-    context.fillStyle = "black";
-    // context.strokeStyle="red"
+  // //   }, [color]);
+  // useEffect(() => {
+  //   // console.log("iiiiiiii", color);
+  //   canvas = canvasRef.current;
+  //   //Our first draw
+  //   context = canvas.getContext("2d");
+  //   context.fillStyle = "black";
+  //   // context.strokeStyle="red"
 
-    context.fillRect(0, 0, context.canvas.width, context.canvas.height);
+  //   context.fillRect(0, 0, 80, 80);
 
-    canvas.addEventListener("mousedown", onMouseDown, false);
-    canvas.addEventListener("mouseup", onMouseUp, false);
-    canvas.addEventListener("mouseout", onMouseUp, false);
-    canvas.addEventListener("mousemove", throttle(onMouseMove, 10), false);
-    //Touch support for mobile devices
-    canvas.addEventListener("touchstart", onMouseDown, false);
-    canvas.addEventListener("touchend", onMouseUp, false);
-    canvas.addEventListener("touchcancel", onMouseUp, false);
-    canvas.addEventListener("touchmove", throttle(onMouseMove, 10), false);
-  }, []);
+  //   canvas.addEventListener("mousedown", onMouseDown, false);
+  //   canvas.addEventListener("mouseup", onMouseUp, false);
+  //   canvas.addEventListener("mouseout", onMouseUp, false);
+  //   canvas.addEventListener("mousemove", throttle(onMouseMove, 10), false);
+  //   //Touch support for mobile devices
+  //   canvas.addEventListener("touchstart", onMouseDown, false);
+  //   canvas.addEventListener("touchend", onMouseUp, false);
+  //   canvas.addEventListener("touchcancel", onMouseUp, false);
+  //   canvas.addEventListener("touchmove", throttle(onMouseMove, 10), false);
+  // }, []);
 
-  var drawing = false;
-  var current = {
-    // color: colorss.color,
-  };
+  // var drawing = false;
+  // var current = {
+  //   // color: colorss.color,
+  // };
 
-  //  canvas.addEventListener('mousemove', onMouseDown,false)
-  function onMouseDown(e) {
-    drawing = true;
-    current.x = e.clientX || e.touches[0].clientX;
-    current.y = e.clientY || e.touches[0].clientY;
-  }
+  // //  canvas.addEventListener('mousemove', onMouseDown,false)
+  // function onMouseDown(e) {
+  //   drawing = true;
+  //   current.x = e.clientX || e.touches[0].clientX;
+  //   current.y = e.clientY || e.touches[0].clientY;
+  // }
 
-  function drawLine(x0, y0, x1, y1, color, emit) {
-    context.beginPath();
-    context.moveTo(x0, y0);
-    context.lineTo(x1, y1);
-    context.strokeStyle = color;
-    context.lineWidth = 1;
-    context.stroke();
-    context.closePath();
+  // function drawLine(x0, y0, x1, y1, color, emit) {
+  //   context.beginPath();
+  //   context.moveTo(x0, y0);
+  //   context.lineTo(x1, y1);
+  //   context.strokeStyle = color;
+  //   context.lineWidth = 1;
+  //   context.stroke();
+  //   context.closePath();
 
-    // console.log(color);
+  //   // console.log(color);
 
-    if (!emit) {
-      return;
-    }
-  }
+   
+  // }
 
-  function throttle(callback, delay) {
-    var previousCall = new Date().getTime();
-    return function () {
-      var time = new Date().getTime();
+  // function throttle(callback, delay) {
+  //   var previousCall = new Date().getTime();
+  //   return function () {
+  //     var time = new Date().getTime();
 
-      if (time - previousCall >= delay) {
-        previousCall = time;
-        callback.apply(null, arguments);
-      }
-    };
-  }
-  function onMouseUp(e) {
-    if (!drawing) {
-      return;
-    }
-    drawing = false;
-    drawLine(
-      current.x,
-      current.y,
-      e.clientX || e.touches[0].clientX,
-      e.clientY || e.touches[0].clientY,
-      color,
-      true
-    );
-    // console.log("color====", e.target.color);
-  }
-  function onMouseMove(e) {
-    if (!drawing) {
-      return;
-    }
-    drawLine(
-      current.x,
-      current.y,
-      e.clientX || e.touches[0].clientX,
-      e.clientY || e.touches[0].clientY,
-      color,
-      true
-    );
-    // console.log(color);
-    current.x = e.clientX || e.touches[0].clientX;
-    current.y = e.clientY || e.touches[0].clientY;
-  }
+  //     if (time - previousCall >= delay) {
+  //       previousCall = time;
+  //       callback.apply(null, arguments);
+  //     }
+  //   };
+  // }
+  // function onMouseUp(e) {
+  //   if (!drawing) {
+  //     return;
+  //   }
+  //   drawing = false;
+  //   drawLine(
+  //     current.x,
+  //     current.y,
+  //     e.clientX || e.touches[0].clientX,
+  //     e.clientY || e.touches[0].clientY,
+  //     color,
+  //     true
+  //   );
+  //   // console.log("color====", e.target.color);
+  // }
+  // function onMouseMove(e) {
+  //   if (!drawing) {
+  //     return;
+  //   }
+  //   drawLine(
+  //     current.x,
+  //     current.y,
+  //     e.clientX || e.touches[0].clientX,
+  //     e.clientY || e.touches[0].clientY,
+  //     color,
+  //     true
+  //   );
+  //   // console.log(color);
+  //   current.x = e.clientX || e.touches[0].clientX;
+  //   current.y = e.clientY || e.touches[0].clientY;
+  // }
 
   return (
     <>
@@ -124,16 +123,19 @@ function DrawIt(props) {
             className=''
             style={{
              
-              background: "whitesmoke",
+              // background: "red",
               
-              alignItems: "center",
-              overflow: "hidden",
+              // alignItems: "center",
+              // overflow: "hidden",
             }}>
-            <canvas
+            {/* <canvas
               ref={canvasRef}
               className='board'
               width={1024}
-              height={622}></canvas>
+              height={622}></canvas> */}
+              <p style={{textAlign:"center",fontSize:"18px"}}>Draw It</p>
+              <p style={{textAlign:"center"}}>Draw on the blackboard using your mouse and bring the creativity here</p>
+              <Canvas/>
           </div>
           <div>
             <h1>Play Details - DrawIt</h1>
